- support for remembering devices and refresh_tokens - async data loading - caching of raw data - add zoom and scroll - add heating_demand_percentage

+++ hive.py
@@ -1,3 +1,5 @@
+import asyncio
+from contextlib import contextmanager
 from datetime import datetime, timedelta
 import hashlib
 from pathlib import Path
@@ -5,20 +7,20 @@ from urllib.parse import urlparse
 import time
 from typing import Any, Callable
 import json
-from pycognito import Cognito
-from pycognito.exceptions import SMSMFAChallengeException
+from aiohttp import ClientResponseError, ClientSession
+from cognito import Cognito, MFAChallengeException
 import pandas as pd
-import requests
 import pydantic
 
-
+SESSION: ClientSession = None
 DATA_DIR = Path("data")
 
 if not DATA_DIR.exists():
     DATA_DIR.mkdir()
+    (DATA_DIR / "raw").mkdir()
 
 CACHED_TOKENS_FILE = DATA_DIR / "tokens.json"
-
+DATA_OVERLAP_DELTA = timedelta(minutes=15)
 
 
 class Credentials(pydantic.BaseModel):
@@ -32,6 +34,9 @@ class AuthenticationResult(pydantic.BaseModel):
     access_token: str | None = None
     refresh_token: str | None = None
     mfa_tokens: dict | None = None
+    mfa_username: str | None = None
+    device_key: str | None = None
+    device_group_key: str | None = None
 
 
 class Device(pydantic.BaseModel):
@@ -48,6 +53,12 @@ class Device(pydantic.BaseModel):
         return self.type == "boilermodule"
 
 
+def _init_session() -> ClientSession:
+    global SESSION
+    SESSION = ClientSession(raise_for_status=True)
+    return SESSION
+
+
 def _load_cached_tokens() -> AuthenticationResult:
     return (AuthenticationResult.model_validate_json(CACHED_TOKENS_FILE.read_text()) if CACHED_TOKENS_FILE.exists() else
             AuthenticationResult())
@@ -59,50 +70,67 @@ def _load_credentials() -> Credentials:
 
 
 def _create_cognito(username: str | None = None, id_token: str | None = None, access_token: str | None = None,
-                    refresh_token: str | None = None):
+                    refresh_token: str | None = None, device_key: str | None = None, device_group_key: str | None = None) -> Cognito:
     return Cognito(user_pool_id='eu-west-1_SamNfoWtf', client_id='3rl4i0ajrmtdm8sbre54p9dvd9',
-                   username=username, id_token=id_token, access_token=access_token, refresh_token=refresh_token)
+                   username=username, id_token=id_token, access_token=access_token, refresh_token=refresh_token,
+                   device_key=device_key, device_group_key=device_group_key)
 
 
-def _timer() -> Callable:
+@contextmanager
+def time_it(msg: str) -> Any:
     start = time.time()
-    return lambda: int((time.time() - start)*1000)
+    try:
+        yield start
+    finally:
+        print(f"{msg} – {int((time.time() - start)*1000)}ms")
+
+def timeit(func):
+    def timed(*args, **kwargs):
+        start = time.time()
+        try:
+            return func(*args, **kwargs)
+        finally:
+            print(f"{func.__name__} – {int((time.time() - start)*1000)}ms")
+    return timed
 
 
+@timeit
 def start_authentication(username: str, password: str) -> AuthenticationResult:
     assert username
     assert password
 
-    t = _timer()
     u = _create_cognito(username=username)
 
     try:
-        print("Invoke authenticate()")
         u.authenticate(password)
-    except SMSMFAChallengeException as mfa_challenge:
-        print(f" done [{t()}ms] *** SMS MFA required")
-        return AuthenticationResult(mfa_tokens=mfa_challenge.get_tokens())
+    except MFAChallengeException as mfa_challenge:
+        print(f" *** SMS MFA required")
+        return AuthenticationResult(mfa_tokens=mfa_challenge.get_tokens(), mfa_username=u.mfa_username)
     else:
         return AuthenticationResult(id_token=u.id_token, access_token=u.access_token, refresh_token=u.refresh_token)
 
 
+@timeit
 def verify_authentication(tokens: AuthenticationResult) -> AuthenticationResult:
-    print("Invoke check_token()")
-    u = _create_cognito(id_token=tokens.id_token, access_token=tokens.access_token, refresh_token=tokens.refresh_token)
+    u = _create_cognito(id_token=tokens.id_token, access_token=tokens.access_token, refresh_token=tokens.refresh_token,
+                        device_key=tokens.device_key, device_group_key=tokens.device_group_key)
     u.check_token()
-    return AuthenticationResult(id_token=u.id_token, access_token=u.access_token, refresh_token=u.refresh_token)
+    return AuthenticationResult(id_token=u.id_token, access_token=u.access_token, refresh_token=u.refresh_token,
+                                device_key=tokens.device_key, device_group_key=tokens.device_group_key)
 
 
+@timeit
 def complete_authentication(username: str, mfa_code: str, mfa_tokens: Any) -> AuthenticationResult:
     assert username
     assert mfa_code
     assert mfa_tokens
-    t = _timer()
+    
     u = _create_cognito(username=username)
-    print("Invoke respond_to_sms_mfa_challenge()")
     u.respond_to_sms_mfa_challenge(mfa_code, mfa_tokens)
-    print(f"  done [{t()}ms]")
-    return AuthenticationResult(id_token=u.id_token, access_token=u.access_token, refresh_token=u.refresh_token)
+    u.confirm_device()
+    u.update_device_status(True)
+    return AuthenticationResult(id_token=u.id_token, access_token=u.access_token, refresh_token=u.refresh_token,
+                                device_key=u.device_key, device_group_key=u.device_group_key)
 
 
 def authenticate(credentials: Credentials) -> AuthenticationResult:
@@ -111,11 +139,11 @@ def authenticate(credentials: Credentials) -> AuthenticationResult:
     try:
         tokens = (verify_authentication(tokens) if tokens.id_token else
                   start_authentication(credentials.username, credentials.password) if not tokens.mfa_tokens else
-                  complete_authentication(credentials.username, credentials.mfa_code, tokens.mfa_tokens))
+                  complete_authentication(tokens.mfa_username, credentials.mfa_code, tokens.mfa_tokens))
 
         CACHED_TOKENS_FILE.write_text(tokens.model_dump_json())
     except Exception:
-        CACHED_TOKENS_FILE.unlink()
+        (CACHED_TOKENS_FILE.unlink() if CACHED_TOKENS_FILE.exists() else None)
         raise
 
     return tokens
@@ -124,7 +152,6 @@ def authenticate(credentials: Credentials) -> AuthenticationResult:
 def get_auth_token() -> str:
     tokens = _load_cached_tokens()
     assert tokens.id_token
-    tokens = authenticate(None)
     return tokens.id_token
 
 
@@ -132,47 +159,64 @@ def md5_hash(input: str) -> str:
     return hashlib.md5(input.encode()).hexdigest()
 
 
-def get_measurements(device: str, start: datetime, end: datetime) -> tuple[str, dict | None]:
+async def get_measurements(device: str, start: datetime, end: datetime) -> tuple[str, dict | None]:
+    # print(locals())
     params = {"from": int(start.timestamp()),
               "to": int(end.timestamp())}
-    return call_api(f"https://measurements.tsdb.prod.bgchprod.info/device/{device}", params=params)
+    return await call_api_async(f"https://measurements.tsdb.prod.bgchprod.info/device/{device}", params=params)
 
 
-def get_devices() -> dict[str, Device]:
+async def get_devices() -> dict[str, Device]:
     file = Path(f"data/devices.json")
 
     if file.exists():
         data = json.loads(file.read_text())
     else:
-        data = call_api("https://beekeeper-uk.hivehome.com/1.0/devices")
+        data = await call_api_async("https://beekeeper-uk.hivehome.com/1.0/devices")
         file.write_text(json.dumps(data, indent=2))
 
     return {x["id"]: Device(id=x["id"], type=x["type"], name=x["state"]["name"])
             for x in data}
 
 
-def fetch_device_data(devices: list[str], start_date: datetime | None) -> dict:
-    end_date = datetime.now()
-    start_date = start_date or datetime(2023, 3, 1, 0, 0, 0)
+async def fetch_device_data(device: str, return_full_history: bool) -> tuple[str, dict]:
+    history_file = Path(f"data/raw/{device}.json")
+    if history_file.exists():
+        old_data = json.loads(history_file.read_text())
+        start_time = pd.to_datetime(old_data["period"]["to"], unit='s', utc=True).to_pydatetime() - DATA_OVERLAP_DELTA
+    else:
+        old_data, start_time = None, datetime(2023, 3, 1, 0, 0, 0)
 
-    print(f"{start_date}({start_date.timestamp()}) / {end_date}({end_date.timestamp()})")
-    device_data = {}
-    for device in devices:
-        try:
-            device_data[device] = get_measurements(device, start_date, end_date)
-        except requests.HTTPError as e:
-            if e.response.status_code == 404:
-                continue
+    try:
+        new_data = await get_measurements(device, start_time, datetime.now())
+        history_data = merge_device_data(old_data, new_data)
+        history_file.write_text(json.dumps(history_data))
+        return (device, history_data if return_full_history else new_data)
+    except ClientResponseError as e:
+        if e.status != 404:
             raise
 
-    Path("data/raw/last_device_data.json").write_text(json.dumps(device_data))
-    return device_data
+
+def merge_device_data(data: dict | None, new_data: dict) -> dict:
+    if not data:
+        return new_data
+
+    max_time_entries = []
+    for key, time_entries in new_data.items():
+        if key not in ["id", "period"]:
+            data.setdefault(key, {}).update(time_entries)
+            max_time_entries.append(max(time_entries.keys()))
+
+    data["period"]["to"] = max(max_time_entries)
+
+    return data
 
 
 def create_device_dataframe(devices: dict[str, Device], data: dict) -> pd.DataFrame:
     device_names = {x.id: x.name for x in devices.values()}
-    df = (pd.concat((pd.concat(pd.DataFrame(device_data.get(m, {}).items(), columns=["date", "value"]).assign(measure=m)
-                               for m in ["heat_target", "heating_relay", "temperature", "heating_demand"])
+    df = (pd.concat((pd.concat(pd.DataFrame(mv.items(), columns=["date", "value"]).assign(measure=m)
+                               for m, mv in device_data.items()
+                               if mv and m in ["heat_target", "heating_relay", "temperature", "heating_demand", "heating_demand_percentage"])
                        .assign(date=lambda x: pd.to_datetime(x.date.astype(int), unit='s', utc=True).dt.floor('T'))
                        .assign(device_id=device_id,
                                value=lambda x: x.value.where((x.measure != "heating_relay") | devices[device_id].is_heater, None))
@@ -189,75 +233,89 @@ def create_device_dataframe(devices: dict[str, Device], data: dict) -> pd.DataFr
             .sort_values(["date", "device_name"]))
 
 
-def get_product_state() -> dict[str, dict]:
-    products = fetch_products()
+async def get_product_state() -> dict[str, dict]:
+    products = await fetch_products()
     result = {x["props"]["trvs"][0]: x for x in products if x["type"] == "trvcontrol"}
     heating = next(x for x in products if x["type"] == "heating")
     result[heating["props"]["zone"]] = heating
     return result
 
 
-def fetch_products():
-    products = call_api("https://beekeeper.hivehome.com/1.0/products")
+async def fetch_products():
+    products = await call_api_async("https://beekeeper.hivehome.com/1.0/products")
     Path("data/products.json").write_text(json.dumps(products, indent=2))
     return products
 
 
-def get_current_device_data(devices: dict[str, Device]) -> pd.DataFrame:
+async def get_current_device_data(devices: dict[str, Device]) -> pd.DataFrame:
+    product_state = await get_product_state()
     return (pd.DataFrame(dict(date=pd.to_datetime(datetime.utcnow(), utc=True).floor('T'),
                               device_id=k,
                               device_name=devices[k].name,
                               temperature=v["props"]["temperature"],
                               heat_target=v["state"]["target"],
                               heating_relay=(v["props"]["working"] if v["type"] == "heating" else None))
-                         for k, v in get_product_state().items())
+                         for k, v in product_state.items())
             .assign(heating_relay=lambda x: x.heating_relay.astype('boolean')))
 
 
-def get_device_data(refresh: bool = False) -> pd.DataFrame:
+async def get_device_data(refresh: bool = False) -> pd.DataFrame:
     try:
-        last_df = pd.read_pickle("data/device_data.pickle")
+        history_data, fetch_partial_data = pd.read_pickle("data/device_data.pickle"), True
         if not refresh:
-            return last_df
-        batch_start_date = last_df.date.max().to_pydatetime() - timedelta(minutes=15)
+            return history_data
     except FileNotFoundError:
-        last_df, batch_start_date = None, None
+        history_data, fetch_partial_data = None, False
 
-    t = _timer()
-    devices = get_devices()
-    batch_df = pd.concat([create_device_dataframe(devices, fetch_device_data(devices.keys(), batch_start_date)),
-                          get_current_device_data(devices)], ignore_index=True)
-    print(f"Loading device data: {t()}ms")
-    t = _timer()
+    with time_it("Loading device data"):
+        devices = await get_devices()
+        async with asyncio.TaskGroup() as tg:
+            current_task = tg.create_task(get_current_device_data(devices))
+            devices_tasks = [tg.create_task(fetch_device_data(device_id, not fetch_partial_data))
+                            for device_id, device in devices.items()
+                            if device.is_heater or device.is_trv]
 
-    new_df = pd.concat([last_df.query("date < @batch_start_date"), batch_df], ignore_index=True) if last_df is not None else batch_df
+    device_data_dict = dict([t.result() for t in devices_tasks if t.result()])
+    if fetch_partial_data:
+        Path("data/raw/last_device_data.json").write_text(json.dumps(device_data_dict))
 
-    ordinal_cols = ["heating_relay", "heating_demand", "heat_target"]
-    new_df[ordinal_cols] = new_df[ordinal_cols].fillna(new_df.groupby("device_id")[ordinal_cols].ffill())
+    with time_it("Creating dataframe"):
+        device_data = create_device_dataframe(devices, device_data_dict)
+        if fetch_partial_data:
+            split_date = max(device_data.groupby("device_id").date.first())
+            device_data = pd.concat([history_data.query("date < @split_date"), device_data.query("date >= @split_date")], ignore_index=True)
+        new_df = pd.concat([device_data, current_task.result()], ignore_index=True)
 
-    assert new_df["heating_relay"].dtype == "boolean"
-    assert new_df["heating_demand"].dtype == "boolean"
+    with time_it("Processing device data"):
+        ordinal_cols = ["heating_relay", "heating_demand", "heat_target", "heating_demand_percentage"]
+        new_df[ordinal_cols] = new_df[ordinal_cols].fillna(new_df.groupby("device_id")[ordinal_cols].ffill())
 
-    new_df["temperature"] = new_df.groupby("device_id", group_keys=False)["temperature"].apply(lambda x: x.interpolate())
-    new_df["temperature"] = new_df["temperature"].fillna(new_df.groupby("device_id")["temperature"].bfill())
+        assert new_df["heating_relay"].dtype == "boolean"
+        assert new_df["heating_demand"].dtype == "boolean"
 
-    _add_heating_stats(new_df, new_df["device_id"] == _get_heater_id(devices))
+        new_df["temperature"] = new_df.groupby("device_id", group_keys=False)["temperature"].apply(lambda x: x.interpolate())
+        new_df["temperature"] = new_df["temperature"].fillna(new_df.groupby("device_id")["temperature"].bfill())
 
-    print(f"Processing device data: {t()}ms")
+        _add_heating_stats(new_df, new_df["device_id"] == _get_heater_id(devices))
 
     new_df.to_pickle("data/device_data.pickle")
     return new_df
 
 
-def call_api(url: str, method: str = "get", params: dict | None = None) -> dict:
-    print(f"Invoke {method.upper()} {urlparse(url).path}")
-    t = _timer()
-    r = requests.request(method, url, params=params, headers={"Authorization": get_auth_token()})
-    r.raise_for_status()
-    result = r.json()
-    print(f"  done [{t()}ms]")
-    return result
+async def call_api_async(url: str, method: str = "get", params: dict | None = None) -> dict:
+    async def _request():
+        assert SESSION
+        async with SESSION.request(method, url, params=params, headers={"Authorization": get_auth_token()}) as r:
+            return await r.json()
 
+    with time_it(f"> {method.upper()} {urlparse(url).path}"):
+        try:
+            return await _request()
+        except ClientResponseError as e:
+            if e.status == 401:
+                authenticate(None)
+                return await _request()
+            raise
 
 def _add_heating_stats(df: pd.DataFrame, source_mask: pd.Series = None):
     source_mask = source_mask if source_mask is not None else slice(None)
